+++ cryptographic-message-syntax/src/signing.rs
@@ -142,12 +142,29 @@ impl<'a> SignerBuilder<'a> {
     }
 }
 
+/// Encapsulated content to sign.
+enum SignedContent {
+    /// No content is being signed.
+    None,
+
+    /// Signed content to be embedded in the signature.
+    Inline(Vec<u8>),
+
+    /// Signed content whose digest is to be captured but won't be included in the signature.
+    ///
+    /// Internal value is the raw content, not the digest.
+    External(Vec<u8>),
+}
+
 /// Entity for incrementally deriving a SignedData primitive.
 ///
 /// Use this type for generating an RFC 5652 payload for signed data.
+///
+/// By default, the encapsulated content to sign is empty. Call [Self::content_inline()]
+/// or [Self::content_external()] to define encapsulated content.
 pub struct SignedDataBuilder<'a> {
     /// Encapsulated content to sign.
-    signed_content: Option<Vec<u8>>,
+    signed_content: SignedContent,
 
     /// Entities who will generated signatures.
     signers: Vec<SignerBuilder<'a>>,
@@ -166,7 +183,7 @@ pub struct SignedDataBuilder<'a> {
 impl<'a> Default for SignedDataBuilder<'a> {
     fn default() -> Self {
         Self {
-            signed_content: None,
+            signed_content: SignedContent::None,
             signers: vec![],
             certificates: vec![],
             content_type: Oid(OID_ID_SIGNED_DATA.as_ref().into()),
@@ -175,12 +192,21 @@ impl<'a> Default for SignedDataBuilder<'a> {
 }
 
 impl<'a> SignedDataBuilder<'a> {
-    /// Define the content to sign.
-    ///
-    /// This content will be embedded in the generated payload.
+    /// Define encapsulated content that will be stored inline in the produced signature.
     #[must_use]
-    pub fn signed_content(mut self, data: Vec<u8>) -> Self {
-        self.signed_content = Some(data);
+    pub fn content_inline(mut self, content: Vec<u8>) -> Self {
+        self.signed_content = SignedContent::Inline(content);
+        self
+    }
+
+    /// Define encapsulated content that won't be present in the produced signature.
+    ///
+    /// The content will be digested and that digest conveyed in the built signature.
+    /// But the content itself won't be present in the signature. RFC 5652 refers to
+    /// this as an _external signature_.
+    #[must_use]
+    pub fn content_external(mut self, content: Vec<u8>) -> Self {
+        self.signed_content = SignedContent::External(content);
         self
     }
 
@@ -262,9 +288,15 @@ impl<'a> SignedDataBuilder<'a> {
             let mut hasher = signer.digest_algorithm.digester();
             if let Some(content) = &signer.message_id_content {
                 hasher.update(content);
-            } else if let Some(content) = &self.signed_content {
-                hasher.update(content);
+            } else {
+                match &self.signed_content {
+                    SignedContent::None => {}
+                    SignedContent::Inline(content) | SignedContent::External(content) => {
+                        hasher.update(content)
+                    }
+                }
             }
+            let digest = hasher.finish();
 
             let mut signed_attributes = SignedAttributes::default();
 
@@ -277,11 +309,12 @@ impl<'a> SignedDataBuilder<'a> {
                 ))],
             });
 
+            // Set `messageDigest` field
             signed_attributes.push(Attribute {
                 typ: Oid(Bytes::copy_from_slice(OID_MESSAGE_DIGEST.as_ref())),
                 values: vec![AttributeValue::new(Captured::from_values(
                     Mode::Der,
-                    hasher.finish().as_ref().encode(),
+                    digest.as_ref().encode(),
                 ))],
             });
 
@@ -391,10 +424,12 @@ impl<'a> SignedDataBuilder<'a> {
             digest_algorithms,
             content_info: EncapsulatedContentInfo {
                 content_type: self.content_type.clone(),
-                content: self
-                    .signed_content
-                    .as_ref()
-                    .map(|content| OctetString::new(Bytes::copy_from_slice(content))),
+                content: match &self.signed_content {
+                    SignedContent::None | SignedContent::External(_) => None,
+                    SignedContent::Inline(content) => {
+                        Some(OctetString::new(Bytes::copy_from_slice(content)))
+                    }
+                },
             },
             certificates: if certificates.is_empty() {
                 None
@@ -425,14 +460,14 @@ mod tests {
     const DIGICERT_TIMESTAMP_URL: &str = "http://timestamp.digicert.com";
 
     #[test]
-    fn simple_rsa_signature() {
+    fn simple_rsa_signature_inline() {
         let key = rsa_private_key();
         let cert = rsa_cert();
 
         let signer = SignerBuilder::new(&key, cert);
 
         let ber = SignedDataBuilder::default()
-            .signed_content(vec![42])
+            .content_inline(vec![42])
             .signer(signer)
             .build_der()
             .unwrap();
@@ -452,6 +487,31 @@ mod tests {
     }
 
     #[test]
+    fn simple_rsa_signature_external() {
+        let key = rsa_private_key();
+        let cert = rsa_cert();
+
+        let signer = SignerBuilder::new(&key, cert);
+
+        let ber = SignedDataBuilder::default()
+            .content_external(vec![42])
+            .signer(signer)
+            .build_der()
+            .unwrap();
+
+        let signed_data = crate::SignedData::parse_ber(&ber).unwrap();
+        assert!(signed_data.signed_content().is_none());
+
+        for signer in signed_data.signers() {
+            signer.verify_message_digest_with_content(&[42]).unwrap();
+            signer
+                .verify_signature_with_signed_data(&signed_data)
+                .unwrap();
+            assert!(signer.unsigned_attributes.is_none());
+        }
+    }
+
+    #[test]
     fn time_stamp_url() {
         let key = rsa_private_key();
         let cert = rsa_cert();
@@ -461,7 +521,7 @@ mod tests {
             .unwrap();
 
         let ber = SignedDataBuilder::default()
-            .signed_content(vec![42])
+            .content_inline(vec![42])
             .signer(signer)
             .build_der()
             .unwrap();
@@ -493,7 +553,7 @@ mod tests {
             let (cert, key) = self_signed_ecdsa_key_pair(Some(*curve));
 
             let cms = SignedDataBuilder::default()
-                .signed_content("hello world".as_bytes().to_vec())
+                .content_inline("hello world".as_bytes().to_vec())
                 .certificate(cert.clone())
                 .signer(SignerBuilder::new(&key, cert))
                 .build_der()
@@ -514,7 +574,7 @@ mod tests {
         let (cert, key) = self_signed_ed25519_key_pair();
 
         let cms = SignedDataBuilder::default()
-            .signed_content("hello world".as_bytes().to_vec())
+            .content_inline("hello world".as_bytes().to_vec())
             .certificate(cert.clone())
             .signer(SignerBuilder::new(&key, cert))
             .build_der()